Add link to Github repository in package.json
Adding this little snippet will allow npmjs.org (https://www.npmjs.com/package/moniker) to show the README.md content inline and link back to the project on Github. There are other projects that have similar function and without this backlink it's hard to tell if that npm package refers to this project.

I've also taken the liberty to bump the version so all you have to do is run `npm publish` after merging this change.